Refactor: removed unnecessary files
