+++ src/screens/body.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Outlet } from "react-router-dom";
+
+
+export default function Body() {
+    return (
+        <div className="body">
+            <Outlet />
+        </div>
+    );
+};
