+++ src/screens/characterScreen.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+
+export default function CharacterScreen() {
+    return (
+        <div className="characterScreen">
+            <h2>Characters</h2>
+        </div>
+    );
+};
